feat: 쪽지시험 응시하기 페이지 경고창 컴포넌트 분리
